+++ python/src/algorithms.rs
@@ -71,10 +71,10 @@ pub(crate) fn all_local_reciprocity(g: &PyGraphView) -> Vec<(u64, f64)> {
 
 #[pyfunction]
 pub(crate) fn triplet_count(g: &PyGraphView) -> usize {
-    docbrown_db::program::algo::triplet_count(&g.graph)
+    docbrown_db::algorithms::triplet_count::triplet_count(&g.graph)
 }
 
 #[pyfunction]
 pub(crate) fn global_clustering_coefficient(g: &PyGraphView) -> f64 {
-    docbrown_db::program::algo::clustering_coefficient(&g.graph)
+    docbrown_db::algorithms::clustering_coefficient::clustering_coefficient(&g.graph)
 }
